Move import file context class
