+++ .ci/behat.community.yml
@@ -70,7 +70,7 @@ legacy:
                     - 1024
                 - Pim\Behat\Context\JobContext:
                     - 'Context\FeatureContext'
-                - Pim\Behat\Context\ImportFileContext:
+                - Pim\Behat\Context\Domain\Collect\ImportFileContext:
                     - 'Context\FeatureContext'
                 - Pim\Behat\Context\Storage\FileInfoStorage:
                     - 'Context\FeatureContext'
