+++ .ci/behat.enterprise.yml
@@ -81,7 +81,7 @@ default:
                     - 1024
                 - PimEnterprise\Behat\Context\JobContext:
                     - 'Context\EnterpriseFeatureContext'
-                - Pim\Behat\Context\ImportFileContext:
+                - Pim\Behat\Context\Domain\Collect\ImportFileContext:
                     - 'Context\FeatureContext'
                 - PimEnterprise\Behat\Context\NavigationContext:
                     - 'Context\EnterpriseFeatureContext'
